+++ src/app/code/community/Divante/VueStorefrontIndexer/Api/Mapping/FieldInterface.php
@@ -17,10 +17,10 @@ interface Divante_VueStorefrontIndexer_Api_Mapping_FieldInterface
     const TYPE_BOOLEAN = 'boolean';
     const TYPE_DOUBLE = 'double';
     const TYPE_INTEGER = 'integer';
+    const TYPE_SHORT = 'short';
     const TYPE_LONG = 'long';
     const TYPE_TEXT = 'text';
     const TYPE_DATE = 'date';
-    const TYPE_INT = 'integer';
 
     const DATE_FORMAT = 'yyyy-MM-dd HH:mm:ss||yyyy-MM-dd||epoch_millis';
 }