+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 ## [Unreleased]
 
+## [1.1.0] (2019.08.30)
+
 ### Added
+- Add option to export reviews to ES
+- Add option to export cms pages to ES
+- Set mapping for `url_path` to keyword
 - Allow to use Elasticsearch host without specific port - @cewald (#25)
 - Update `vsf_tools.php` - @cewald (#26)
   - Add `store` option to `reindex` action
@@ -8,7 +13,7 @@
 
 ### Fixed
 - Fixed CMS Block and Page identifier to avoid `-` tokenization - @mtarld (#22)
-- `stock_status` is now taken into consideration for the value of the synchronized value of `is_in_stock` for configurable products, this makes product listings in Vue Storefront better reflect the ones in native Magento - [@indiebytes](https://github.com/indiebytes) [(#24)](https://github.com/DivanteLtd/magento1-vsbridge-indexer/pull/24)
+- Set is_in_stock/stock_status value - base on Stock Avaibility of configurable product and children stock status. `stock_status` value is calculated in inventory indexer and might be out of date. #34
 - Always wrong scheme after enabling HTTPS in backend - @cewald (#25)
 
 ## [1.0.0] (2019.04.09)